pyproject.toml
[tool.poetry]
name = "snow-ipa"
version = "2.2.0"
description = ""
authors = ["Erick G <9361923+silversky54@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
earthengine-api = "^0.1.375"
google-api-python-client = "^2.104.0"
pytz = "^2023.3.post1"
email-validator = "^2.1.0.post1"
pytest-mock = "^3.14.0"
jinja2 = "^3.1.6"
arrow = "^1.3.0"
prettytable = "^3.16.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
jupyter = "^1.0.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = ["--import-mode=importlib"]
testpaths = ["tests"]
